% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{add.alpha}
\usage{
add.alpha(col, alpha = 0.65)
}
\arguments{
\item{col}{color from colors()}

\item{alpha}{alpha level (between 0 and 1)}
}
\description{
A function to add alpha levels to RGB colors.
}
\note{
This code was created by Markus Gesmann. Derek Beaton included the
code in prettyGraphs because it is a versatile function and used in
prettyGraphs. See:
https://magesblog.com/post/2013-04-30-how-to-change-alpha-value-of-colours-in/
}
\author{
Markus Gesmann
}
\keyword{misc}
