\name{prettyGraphs}
\alias{prettyGraphs}
\docType{package}
\title{
prettyGraphs: for publication-quality graphics.
}
\description{
prettyGraphs is a package that produces simple, crisp, publication-quality graphics for multivariate analyses. 
}
\details{
\tabular{ll}{
Package: \tab prettyGraphs\cr
Type: \tab Package\cr
Version: \tab 2.1.5\cr
Date: \tab 2013-12-09\cr
License: \tab GPL-2\cr
Suggests: ExPosition
}
}
\author{Derek Beaton <exposition.software@gmail.com>}
\references{
	Two functions were copied/derived for use in prettyGraphs: peeledHull and repmat.\cr \cr
	
	For peeledHull see:\cr
	http://carme-n.org/?sec=code2\cr
	Greenacre, M. J. (2007). Correspondence Analysis in Practice. \emph{Chapman and Hall}.\cr \cr
	
	For repmat see:\cr
	http://cran.r-project.org/doc/contrib/R-and-octave.txt
}
\keyword{ package }
\seealso{
\code{\link{prettyPlot}}
\code{\link{contributionBars}}
\code{\link{correlationPlotter}}
\code{\link{peeledHull}}
\code{\link{minmaxHelper}}
\code{\link{repmat}}
}
\keyword{ graphs }
\keyword{ multivariate }