\name{prettyGraphsColors}
\alias{prettyGraphsColors}
\title{
prettyGraphsColors
}
\description{
A function to return only a subset of R colors used by the prettyGraphs package.
}
\usage{
prettyGraphsColors()
}
\value{
	\item{possibleColors}{A list of colors used by prettyGraphs}
}

\author{
Derek Beaton
}
\keyword{ misc }