% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyGraphsColorSelection.R
\name{prettyGraphsColorSelection}
\alias{prettyGraphsColorSelection}
\title{Algorithmically select colors from prettyGraphs}
\usage{
prettyGraphsColorSelection(n.colors = 1, offset = NULL, starting.color = 163)
}
\arguments{
\item{n.colors}{number of colors needed}

\item{offset}{numeric. Should be a prime number, if it is not, the closest
prime is selected. This number decides the distance between indices for
color selection.}

\item{starting.color}{numeric. This is the starting location (e.g., color)
in a matrix of \code{\link{prettyGraphsColors}}.}
}
\value{
\item{a matrix}{a matrix of colors are returned.}
}
\description{
This function uses prime numbers to select colors from
\code{\link{prettyGraphsColors}}.
}
\seealso{
\code{\link{prettyGraphsColors}},
\code{\link{prettyGraphsHSVColorSelection}}
}
\author{
Derek Beaton\cr\cr prime number selection from here:\cr
http://stackoverflow.com/questions/3789968/generate-a-list-of-primes-in-r-up-to-a-certain-number
}
\keyword{graphs}
