\name{repmat}
\alias{repmat}
\title{
repmat
}
\description{
A function to match repmat in Matlab (Mathworks, Natick, MA)
}
\usage{
repmat(a,n,m)
}
\arguments{
  \item{a}{
Matrix or vector to repeat.
}
  \item{n}{
Number of row repititions.
}
  \item{m}{
Number of column repititions.
}

}
\references{
	For repmat see:\cr
	http://cran.r-project.org/doc/contrib/R-and-octave.txt
}
\author{
Robin Hankin. 
}
\note{
	This code was created by Robin Hankin. Derek Beaton included the code in prettyGraphs because it is a versatile function and used in prettyGraphs.
	}
\keyword{ misc }