% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyGraphsColors.R
\name{prettyGraphsColors}
\alias{prettyGraphsColors}
\title{prettyGraphsColors}
\usage{
prettyGraphsColors()
}
\value{
\item{possibleColors}{A list of colors used by prettyGraphs}
}
\description{
A function to return only a subset of R colors used by the prettyGraphs
package.
}
\author{
Derek Beaton
}
\keyword{misc}
